TerminalBlocks: Add internal parameter for silk arrow at top
This is a small improvement to the terminal block function make_silk_outline_with_pin1_arrow which will help when a horizontal terminal block has the wires entering at the bottom, and you want the arrow to be unobscured on the in-board side (i.e. not under the entering wires, and less at risk of hitting the board edge).

It also now takes a more semantically meaningful GeomRectangle rather than a point and size (the function predates that class).

See merge request https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1836